Refactor grab controller
